Wait for wineserver shutdown in the RUN commands
Wine is a tricky beast. As it turns out, winetricks uses a wineserver
instance in the background to hold all its processes, and it's a bit
sloppy in tracking which process is still running when it returns. It
resulted in the docker runtime attempting to snapshot a system while
there was sometimes still active processes, and even worse, they could
still be writing on the FS. Hilarity ensues.